fix(safety): force size_t multiplication width on 14 buffer-size constants
bugprone-implicit-widening-of-multiplication-result flags expressions
like `64 * 1024 * 1024` where the multiplication is performed in `int`
but the result is stored in `size_t` / `int64_t` / `uintmax_t`. The
value fits today, but any future scaling tweak (e.g. `256 * 1024 * 1024
* 32`) silently overflows signed-int and is UB.

Two fix patterns applied:

1. Constant multiplications → prefix first operand with `ULL` / `LL`:
     `64 * 1024 * 1024` → `64ULL * 1024 * 1024`
   This forces the entire expression to evaluate in `unsigned long long`.

2. Variable multiplications (e.g. `width_ * height_` where both are
   `int`) → explicit `static_cast<size_t>(width_) * height_` to widen
   the first operand before multiplying.

Sites:
  apps/cli/src/auth_profile_store.cpp      base = 60LL * 1000
  apps/cli/src/cmd/system.cpp              1024ULL * 1024
  apps/cli/src/terminal.cpp        (4×)    static_cast<size_t>(w) * h
  apps/cli/src/tui/workspace.cpp           static_cast<size_t>(depth) * 2
  apps/gateway/include/.../routes.hpp      1ULL * 1024 * 1024
  apps/gateway/src/mcp_fleet_tools.cpp     16ULL * 1024 * 1024
  libs/bridge/src/executor.cpp     (2×)    16ULL/64ULL multiplications
  libs/crypto/src/aes_gcm.cpp              kMaxPlaintextSize 64ULL * 1024 * 1024
  libs/crypto/src/ed25519.cpp              kMaxMessageSize 256ULL * 1024 * 1024
  libs/inference/src/model_registry.cpp    64ULL * 1024 buf_size

Fixes #54


---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co

+++ apps/cli/src/terminal.cpp
@@ -133,8 +133,8 @@ void get_terminal_size(int& w, int& h) {
 
 TerminalScreen::TerminalScreen() {
     get_terminal_size(width_, height_);
-    front_buffer_.assign(width_ * height_, Cell{.ch = 0xFFFFFFFF});
-    back_buffer_.resize(width_ * height_);
+    front_buffer_.assign(static_cast<size_t>(width_) * height_, Cell{.ch = 0xFFFFFFFF});
+    back_buffer_.resize(static_cast<size_t>(width_) * height_);
 }
 
 TerminalScreen::~TerminalScreen() { disable_raw_mode(); }
@@ -142,8 +142,8 @@ TerminalScreen::~TerminalScreen() { disable_raw_mode(); }
 void TerminalScreen::resize(int w, int h) {
     if (w == width_ && h == height_) return;
     width_ = w; height_ = h;
-    front_buffer_.assign(w * h, Cell{.ch = 0xFFFFFFFF});
-    back_buffer_.assign(w * h, Cell{});
+    front_buffer_.assign(static_cast<size_t>(w) * h, Cell{.ch = 0xFFFFFFFF});
+    back_buffer_.assign(static_cast<size_t>(w) * h, Cell{});
 }
 
 void TerminalScreen::clear() { for (auto& c : back_buffer_) c = Cell{}; }
